Read and record current kernel page table root in `TaskContext::new()`
This modification is used to support the initialization of page table root
for kernel threads, but only for kernel threads created at initialization time.

After the user program starts running, the user page table is used for both user
mode running and kernel mode syscalls.

This means that if the user program creates a kernel thread in a syscall, it will
still use the user page table to initialize `TaskContext`, potentially resulting
in undefined behavior.

This unresolved issue is left for later.